+++ lib/cookie.js
@@ -36,6 +36,7 @@ const pubsuffix = require("./pubsuffix-psl");
 const Store = require("./store").Store;
 const MemoryCookieStore = require("./memstore").MemoryCookieStore;
 const pathMatch = require("./pathMatch").pathMatch;
+const validators = require("./validators.js");
 const VERSION = require("./version");
 const { fromCallback } = require("universalify");
 
@@ -78,64 +79,8 @@ const MIN_TIME = 0; // 31-bit min
 const SAME_SITE_CONTEXT_VAL_ERR =
   'Invalid sameSiteContext option for getCookies(); expected one of "strict", "lax", or "none"';
 
-/* Validation functions copied from check-types package - https://www.npmjs.com/package/check-types */
-function isFunction(data) {
-  return typeof data === 'function';
-}
-
-function isNonEmptyString(data) {
-  return isString(data) && data !== '';
-}
-
-function isDate(data) {
-  return isInstanceStrict(data, Date) && isInteger(data.getTime());
-}
-
-function isEmptyString(data) {
-  return data === '';
-}
-
-function isString(data) {
-  return typeof data === 'string';
-}
-
-function isObject(data) {
-  return toString.call(data) === '[object Object]';
-}
-function isInstanceStrict(data, prototype) {
-  try {
-    return data instanceof prototype;
-  } catch (error) {
-    return false;
-  }
-}
-
-function isInteger(data) {
-  return typeof data === 'number' && data % 1 === 0;
-}
-
-
-/* End validation functions */
-
-
-
-function validate(bool, cb, options) {
-  if (!isFunction(cb)) {
-    options = cb;
-    cb = null;
-  }
-  if (!isObject(options)) options = { Error: "Failed Check" };
-  if (!bool) {
-    if (cb) {
-      cb(new ParameterError(options));
-    } else {
-      throw new ParameterError(options);
-    }
-  }
-}
-
 function checkSameSiteContext(value) {
-  validate(isNonEmptyString(value), value);
+  validators.validate(validators.isNonEmptyString(value), value);
   const context = String(value).toLowerCase();
   if (context === "none" || context === "lax" || context === "strict") {
     return context;
@@ -358,7 +303,7 @@ function parseDate(str) {
 }
 
 function formatDate(date) {
-  validate(isDate(date), date);
+  validators.validate(validators.isDate(date), date);
   return date.toUTCString();
 }
 
@@ -461,7 +406,7 @@ function defaultPath(path) {
 }
 
 function trimTerminator(str) {
-  if (isEmptyString(str)) return str;
+  if (validators.isEmptyString(str)) return str;
   for (let t = 0; t < TERMINATORS.length; t++) {
     const terminatorIdx = str.indexOf(TERMINATORS[t]);
     if (terminatorIdx !== -1) {
@@ -474,7 +419,7 @@ function trimTerminator(str) {
 
 function parseCookiePair(cookiePair, looseMode) {
   cookiePair = trimTerminator(cookiePair);
-  validate(isString(cookiePair), cookiePair);
+  validators.validate(validators.isString(cookiePair), cookiePair);
 
   let firstEq = cookiePair.indexOf("=");
   if (looseMode) {
@@ -676,7 +621,7 @@ function parse(str, options) {
  * @returns boolean
  */
 function isSecurePrefixConditionMet(cookie) {
-  validate(isObject(cookie), cookie);
+  validators.validate(validators.isObject(cookie), cookie);
   return !cookie.key.startsWith("__Secure-") || cookie.secure;
 }
 
@@ -692,7 +637,7 @@ function isSecurePrefixConditionMet(cookie) {
  * @returns boolean
  */
 function isHostPrefixConditionMet(cookie) {
-  validate(isObject(cookie));
+  validators.validate(validators.isObject(cookie));
   return (
     !cookie.key.startsWith("__Host-") ||
     (cookie.secure &&
@@ -760,8 +705,8 @@ function fromJSON(str) {
  */
 
 function cookieCompare(a, b) {
-  validate(isObject(a), a);
-  validate(isObject(b), b);
+  validators.validate(validators.isObject(a), a);
+  validators.validate(validators.isObject(b), b);
   let cmp = 0;
 
   // descending for length: b CMP a
@@ -789,7 +734,7 @@ function cookieCompare(a, b) {
 // Gives the permutation of all possible pathMatch()es of a given path. The
 // array is in longest-to-shortest order.  Handy for indexing.
 function permutePath(path) {
-  validate(isString(path));
+  validators.validate(validators.isString(path));
   if (path === "/") {
     return ["/"];
   }
@@ -838,12 +783,6 @@ const cookieDefaults = {
   creation: null,
   lastAccessed: null,
   sameSite: "none"
-};
-
-class ParameterError extends Error {
-  constructor(...params) {
-    super(...params);
-  }
 };
 
 class Cookie {
@@ -1131,7 +1070,7 @@ class CookieJar {
     if (typeof options === "boolean") {
       options = { rejectPublicSuffixes: options };
     }
-    validate(isObject(options), options);
+    validators.validate(validators.isObject(options), options);
     this.rejectPublicSuffixes = options.rejectPublicSuffixes;
     this.enableLooseMode = !!options.looseMode;
     this.allowSpecialUseDomain = !!options.allowSpecialUseDomain;
@@ -1148,20 +1087,20 @@ class CookieJar {
   }
 
   setCookie(cookie, url, options, cb) {
-    validate(isNonEmptyString(url), cb, options);   
+    validators.validate(validators.isNonEmptyString(url), cb, options);   
     let err;
 
-    if (isFunction(url)) {
+    if (validators.isFunction(url)) {
       cb = url;
       return cb(new Error("No URL was specified"));
     }
 
     const context = getCookieContext(url);
-    if (isFunction(options)) {
+    if (validators.isFunction(options)) {
       cb = options;
       options = {};
     }
-    validate(isFunction(cb), cb);
+    validators.validate(validators.isFunction(cb), cb);
 
     const host = canonicalDomain(context.hostname);
     const loose = options.loose || this.enableLooseMode;
@@ -1329,14 +1268,14 @@ class CookieJar {
 
   // RFC6365 S5.4
   getCookies(url, options, cb) {
-    validate(isNonEmptyString(url), cb, url);
+    validators.validate(validators.isNonEmptyString(url), cb, url);
     const context = getCookieContext(url);
-    if (isFunction(options)) {
+    if (validators.isFunction(options)) {
       cb = options;
       options = {};
     }
-    validate(isObject(options), cb, options);
-    validate(isFunction(cb), cb);
+    validators.validate(validators.isObject(options), cb, options);
+    validators.validate(validators.isFunction(cb), cb);
 
     const host = canonicalDomain(context.hostname);
     const path = context.pathname || "/";
@@ -1452,7 +1391,7 @@ class CookieJar {
 
   getCookieString(...args) {
     const cb = args.pop();
-    validate(isFunction(cb), cb);
+    validators.validate(validators.isFunction(cb), cb);
     const next = function(err, cookies) {
       if (err) {
         cb(err);
@@ -1472,7 +1411,7 @@ class CookieJar {
 
   getSetCookieStrings(...args) {
     const cb = args.pop();
-    validate(isFunction(cb), cb);
+    validators.validate(validators.isFunction(cb), cb);
     const next = function(err, cookies) {
       if (err) {
         cb(err);
@@ -1490,9 +1429,9 @@ class CookieJar {
   }
 
   serialize(cb) {
-    validate(isFunction(cb), cb);
+    validators.validate(validators.isFunction(cb), cb);
     let type = this.store.constructor.name;
-    if (isObject(type)) {
+    if (validators.isObject(type)) {
       type = null;
     }
 
@@ -1610,7 +1549,7 @@ class CookieJar {
   }
 
   removeAllCookies(cb) {
-    validate(isFunction(cb), cb);
+    validators.validate(validators.isFunction(cb), cb);
     const store = this.store;
 
     // Check that the store implements its own removeAllCookies(). The default
@@ -1664,7 +1603,7 @@ class CookieJar {
       cb = store;
       store = null;
     }
-    validate(isFunction(cb), cb);
+    validators.validate(validators.isFunction(cb), cb);
 
     let serialized;
     if (typeof strOrObj === "string") {
@@ -1757,4 +1696,4 @@ exports.permuteDomain = require("./permuteDomain").permuteDomain;
 exports.permutePath = permutePath;
 exports.canonicalDomain = canonicalDomain;
 exports.PrefixSecurityEnum = PrefixSecurityEnum;
-exports.ParameterError = ParameterError;
+exports.ParameterError = validators.ParameterError;
